Fix delete function

+++ assets/scripts/rounds/events.js
@@ -21,21 +21,21 @@ const onSubmitRound = function (event) {
   let data = getFormFields(event.target);
   data.profile_id = app.profile.id;
 
-  if (app.profile.rounds.length < 15) {
+  // if (app.profile.rounds.length < 14) {
     // console.log('fewer than 15 rounds');
     api.createRound(data)
       .done(ui.createRoundSuccess)
       .fail(ui.createRoundFailure);
-      setTimeout(function(){console.log('rounds are now', app.profile.rounds);}, 500);
+      setTimeout(function(){console.log('rounds are now', app.profile.rounds);}, 100);
     return;
-  }
-  else {
-    // console.log('more than 15 rounds');
-    api.createRound(data)
-      .done(ui.createMaxRoundSuccess)
-      .fail(ui.createRoundFailure);
-      setTimeout(function(){console.log('rounds are now', app.profile.rounds);}, 500);
-  }
+  // }
+  // else {
+  //   // console.log('more than 15 rounds');
+  //   api.createRound(data)
+  //     .done(ui.createMaxRoundSuccess)
+  //     .fail(ui.createRoundFailure);
+  //     setTimeout(function(){console.log('rounds are now', app.profile.rounds);}, 500);
+  // }
 
 };
 //
@@ -58,9 +58,17 @@ const onViewRoundsClick = function (data) {
 // - Flip array back to order matching API before adding new round.
 
 const onDeleteLastRound = function () {
-  let roundID = app.profile.rounds.pop().id;
+  let roundID;
+  if (app.profile.rounds[0] === undefined) {
+    return;
+  }
+  else if (app.profile.rounds.length === 1) {
+    roundID = app.profile.rounds[0].id;
+  } else {
+    roundID = app.profile.rounds.pop().id;
+  }
   api.destroyRound(roundID)
-    .done(ui.deleteRoundSuccess, app.profile.rounds.shift());
+    .done(ui.deleteRoundSuccess);
   ui.deleteLastRoundSuccess();
   setTimeout(function(){console.log('rounds are now', app.profile.rounds);}, 500);
 };
